third_party/blink/web_tests/media/mediasession/mojo/position-state-propagated.html

<!DOCTYPE html>
<title>MediaSession Mojo Test</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/mediasession/media_session.mojom.js"></script>
<script src="resources/mediasessionservice-mock.js"></script>
<script src="resources/utils.js"></script>
<script>
const inputPositions = [
  { duration: 10 },
  { duration: 10, playbackRate: 2.0, position: 5 },
  null
];

const expectations = [
  { duration: 10, playbackRate: 1, position: 0 },
  { duration: 10, playbackRate: 2.0, position: 5 },
  null
];

function toSecs(timeDelta) {
  return timeDelta.microseconds / 1000000;
}

async_test((t) => {
  let m = mediaSessionServiceMock;
  let nextExpectation = 0;

  mediaSessionServiceMock.setPositionStateCallback(t.step_func((position) => {
    const expectation = expectations[nextExpectation++];

    if (expectation === null) {
      assert_equals(expectation, position);
    } else {
      assert_equals(toSecs(position.duration), expectation.duration);
      assert_equals(toSecs(position.position), expectation.position);
      assert_equals(position.playbackRate, expectation.playbackRate);
    }

    if (nextExpectation == inputPositions.length)
      t.done();
  }));

  for (let position of inputPositions)
    window.navigator.mediaSession.setPositionState(position);
}, "test that MediaSession.setPositionState is correctly propagated");
</script>
